backend/requirements.txt
flask
flask-cors
requests
python-dotenv
gunicorn==20.1.0
pandas
numpy
scikit-learn
joblib
openpyxl
xlsxwriter
google-auth